Moved Trap enum declaration into the trap_handlers module.
