+++ src/ppytty/kernel/trap_handlers.py
@@ -15,7 +15,6 @@ from . import common
 from . import loop
 from . state import state
 from . window import Window
-from . trap_ids import Trap
 
 
 
@@ -23,12 +22,44 @@ log = logging.getLogger(__name__)
 
 
 
-# Maps Trap enums to handler functions.
+# ----------------------------------------------------------------------------
+# Trap idenfifiers
+
+class Trap(enum.Enum):
+
+    DIRECT_CLEAR = enum.auto()
+    DIRECT_PRINT = enum.auto()
+
+    WINDOW_CREATE = enum.auto()
+    WINDOW_RENDER = enum.auto()
+    WINDOW_DESTROY = enum.auto()
+
+    SLEEP = enum.auto()
+
+    KEY_READ = enum.auto()
+    KEY_UNREAD = enum.auto()
+
+    TASK_SPAWN = enum.auto()
+    TASK_DESTROY = enum.auto()
+    TASK_WAIT = enum.auto()
+
+    MESSAGE_SEND = enum.auto()
+    MESSAGE_WAIT = enum.auto()
+
+    STATE_DUMP = enum.auto()
+
+
+
+# ----------------------------------------------------------------------------
+# Trap identifier to handler function mapping.
+
+
+# Maps Trap identifiers to handler functions.
 
 handlers = {}
 
 
-# Updates the handlers mapping.
+# Decorator function that populates the handlers map.
 
 def handler_for(trap_id):
 
@@ -42,6 +73,10 @@ def handler_for(trap_id):
 
     return decorator_function
 
+
+
+# ----------------------------------------------------------------------------
+# Trap handlers
 
 
 @handler_for(Trap.DIRECT_CLEAR)
